Tweaks to improve click accuracy
The speed slider was throwing off coordinates so it's disabled for now. Margins were also causing problems in scaled models so they're removed. The click location of the carry tool is now moved to the tips of the fingers. The location of the bubble popups is also slightly tweaked to point directly to the info tool.